add api: deleteServiceColumn, deleteLabel
