+++ s2core/src/main/scala/org/apache/s2graph/core/S2VertexProperty.scala
@@ -48,6 +48,9 @@ case class S2VertexProperty[V](element: S2VertexLike,
     innerVal.bytes
   }
 
+
+  val valueAny = castValue(v, columnMeta.dataType)
+
   val value = castValue(v, columnMeta.dataType).asInstanceOf[V]
 
   override def properties[U](strings: String*): util.Iterator[Property[U]] = ???
