+++ s2core/src/main/scala/org/apache/s2graph/core/mysqls/Service.scala
@@ -123,6 +123,6 @@ case class Service(id: Option[Int],
 
   lazy val extraOptions = Model.extraOptions(options)
   lazy val storageConfigOpt: Option[Config] = toStorageConfig
-  lazy val serviceColumns: Seq[ServiceColumn] = ServiceColumn.findByServiceId(id.get, useCache = true)
+  def serviceColumns(useCache: Boolean): Seq[ServiceColumn] = ServiceColumn.findByServiceId(id.get, useCache = useCache)
   def toStorageConfig: Option[Config] = Model.toStorageConfig(extraOptions)
 }
