+++ s2graphql/src/main/resources/application.conf
@@ -22,12 +22,12 @@ akka {
   loglevel = "INFO"
 }
 
-//db.default.password = sa
-//db.default.user = sa
-//s2graph.storage.backend = rocks
-//rocks.storage.file.path = rocks_db
-//rocks.storage.mode = production
-//rocks.storage.ttl = -1
-//rocks.storage.read.only = false
+db.default.password = sa
+db.default.user = sa
+s2graph.storage.backend = rocks
+rocks.storage.file.path = rocks_db
+rocks.storage.mode = production
+rocks.storage.ttl = -1
+rocks.storage.read.only = false
