+++ s2graphql/src/main/scala/org/apache/s2graph/graphql/marshaller/package.scala
@@ -1,0 +1,102 @@
+package org.apache.s2graph.graphql
+
+import org.apache.s2graph.core.Management.JsonModel._
+import org.apache.s2graph.graphql.types.S2Type._
+import org.apache.s2graph.graphql.types.S2ManagementType._
+import sangria.marshalling._
+
+package object marshaller {
+
+  implicit object IndexFromInput extends FromInput[Index] {
+    val marshaller = CoercedScalaResultMarshaller.default
+
+    def fromResult(node: marshaller.Node) = {
+      val input = node.asInstanceOf[Map[String, Any]]
+      Index(input("name").asInstanceOf[String], input("propNames").asInstanceOf[Seq[String]])
+    }
+  }
+
+  implicit object PropFromInput extends FromInput[Prop] {
+    val marshaller = CoercedScalaResultMarshaller.default
+
+    def fromResult(node: marshaller.Node) = {
+      val input = node.asInstanceOf[Map[String, String]]
+      Prop(input("name"), input("defaultValue"), input("dataType"))
+    }
+  }
+
+  implicit object LabelServiceFromInput extends FromInput[LabelServiceProp] {
+    val marshaller = CoercedScalaResultMarshaller.default
+
+    def fromResult(node: marshaller.Node) = {
+      val input = node.asInstanceOf[Map[String, String]]
+      LabelServiceProp(input("name"), input("columnName"), input("dataType"))
+    }
+  }
+
+  implicit object PartialServiceVertexParamFromInput extends FromInput[Vector[PartialServiceVertexParam]] {
+    val marshaller = CoercedScalaResultMarshaller.default
+
+    def fromResult(node: marshaller.Node) = {
+      val inputMap = node.asInstanceOf[Map[String, marshaller.Node]]
+
+      val ret = inputMap.toVector.map { case (columnName, node) =>
+        val param = PartialVertexFromInput.fromResult(node)
+        PartialServiceVertexParam(columnName, param)
+      }
+
+      ret
+    }
+  }
+
+  implicit object PartialVertexFromInput extends FromInput[PartialVertexParam] {
+    val marshaller = CoercedScalaResultMarshaller.default
+
+    def fromResult(node: marshaller.Node) = {
+
+      val _inputMap = node.asInstanceOf[Option[Map[String, Any]]]
+      val inputMap = _inputMap.get
+
+      val id = inputMap("id")
+      val ts = inputMap.get("timestamp") match {
+        case Some(Some(v)) => v.asInstanceOf[Long]
+        case _ => System.currentTimeMillis()
+      }
+      val props = inputMap.get("props") match {
+        case Some(Some(v)) => v.asInstanceOf[Map[String, Option[Any]]].filter(_._2.isDefined).mapValues(_.get)
+        case _ => Map.empty[String, Any]
+      }
+
+      PartialVertexParam(ts, id, props)
+    }
+  }
+
+  implicit object PartialEdgeFromInput extends FromInput[PartialEdgeParam] {
+    val marshaller = CoercedScalaResultMarshaller.default
+
+    def fromResult(node: marshaller.Node) = {
+      val inputMap = node.asInstanceOf[Map[String, Any]]
+
+      val from = inputMap("from")
+      val to = inputMap("to")
+
+      val ts = inputMap.get("timestamp") match {
+        case Some(Some(v)) => v.asInstanceOf[Long]
+        case _ => System.currentTimeMillis()
+      }
+
+      val dir = inputMap.get("direction") match {
+        case Some(Some(v)) => v.asInstanceOf[String]
+        case _ => "out"
+      }
+
+      val props = inputMap.get("props") match {
+        case Some(Some(v)) => v.asInstanceOf[Map[String, Option[Any]]].filter(_._2.isDefined).mapValues(_.get)
+        case _ => Map.empty[String, Any]
+      }
+
+      PartialEdgeParam(ts, from, to, dir, props)
+    }
+  }
+
+}
